# Optimize generic types