Based on #186, this PR adds a few optimizations to reduce the CPU footprint of.

I think we're getting close to the maximum throughput we can get using just Go, this will be a good baseline to compare to for future optimizations:

```
goos: linux
goarch: amd64
pkg: github.com/segmentio/parquet-go
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
```

```
name                                          old time/op  new time/op  delta
GenericBuffer/benchmarkRowType/go1.18          384ns ± 5%   174ns ± 3%   -54.79%  (p=0.000 n=10+9)
GenericBuffer/booleanColumn/go1.18             464ns ± 4%    97ns ± 4%   -79.06%  (p=0.000 n=9+10)
GenericBuffer/int32Column/go1.18               226ns ± 5%    72ns ± 2%   -68.08%  (p=0.000 n=10+10)
GenericBuffer/int64Column/go1.18               223ns ± 2%    72ns ± 4%   -67.64%  (p=0.000 n=8+10)
GenericBuffer/floatColumn/go1.18               208ns ±14%    77ns ± 3%   -62.90%  (p=0.000 n=10+10)
GenericBuffer/doubleColumn/go1.18              217ns ± 3%    78ns ± 4%   -64.17%  (p=0.000 n=10+10)
GenericBuffer/byteArrayColumn/go1.18          1.26µs ± 2%  1.07µs ± 3%   -14.91%  (p=0.000 n=8+9)
GenericBuffer/fixedLenByteArrayColumn/go1.18   587ns ± 4%   604ns ± 3%    +2.93%  (p=0.001 n=10+9)
GenericBuffer/stringColumn/go1.18             1.03µs ±27%  0.66µs ± 2%   -35.88%  (p=0.000 n=10+8)
GenericBuffer/indexedStringColumn/go1.18      4.25µs ± 3%  4.21µs ± 3%      ~     (p=0.289 n=10+10)
GenericBuffer/uuidColumn/go1.18                180ns ± 2%   108ns ± 2%   -40.13%  (p=0.000 n=10+9)
GenericBuffer/mapColumn/go1.18                72.7µs ± 8%  72.1µs ± 2%      ~     (p=0.400 n=10+9)
GenericBuffer/decimalColumn/go1.18             224ns ± 4%    72ns ± 3%   -67.67%  (p=0.000 n=10+10)
GenericBuffer/contact/go1.18                  1.79µs ± 7%  1.39µs ± 8%   -22.50%  (p=0.000 n=10+10)

name                                          old row/s    new row/s    delta
GenericBuffer/benchmarkRowType/go1.18           260M ± 5%    576M ± 3%  +121.14%  (p=0.000 n=10+9)
GenericBuffer/booleanColumn/go1.18              216M ± 4%   1029M ± 4%  +377.50%  (p=0.000 n=9+10)
GenericBuffer/int32Column/go1.18                443M ± 5%   1386M ± 2%  +213.15%  (p=0.000 n=10+10)
GenericBuffer/int64Column/go1.18                445M ± 6%   1387M ± 4%  +211.54%  (p=0.000 n=9+10)
GenericBuffer/floatColumn/go1.18                484M ±13%   1299M ± 3%  +168.38%  (p=0.000 n=10+10)
GenericBuffer/doubleColumn/go1.18               460M ± 3%   1285M ± 4%  +179.13%  (p=0.000 n=10+10)
GenericBuffer/byteArrayColumn/go1.18           78.9M ± 5%   93.4M ± 3%   +18.31%  (p=0.000 n=9+9)
GenericBuffer/fixedLenByteArrayColumn/go1.18    170M ± 4%    166M ± 3%    -2.85%  (p=0.001 n=10+9)
GenericBuffer/stringColumn/go1.18              99.5M ±23%  152.1M ± 2%   +52.96%  (p=0.000 n=10+8)
GenericBuffer/indexedStringColumn/go1.18       23.5M ± 3%   23.8M ± 3%      ~     (p=0.280 n=10+10)
GenericBuffer/uuidColumn/go1.18                 556M ± 2%    928M ± 2%   +66.99%  (p=0.000 n=10+9)
GenericBuffer/mapColumn/go1.18                 1.38M ± 7%   1.39M ± 2%      ~     (p=0.400 n=10+9)
GenericBuffer/decimalColumn/go1.18              446M ± 4%   1380M ± 3%  +209.24%  (p=0.000 n=10+10)
GenericBuffer/contact/go1.18                   55.8M ± 7%   72.1M ± 7%   +29.02%  (p=0.000 n=10+10)
```

One interesting aspect of this change is that both the code paths using type information and the code using the `parquet.Value` abstractions now use the same underlying implementation. Future improvements will then apply to either approach, yielding greater efficiency in a wider variety of use cases.